move merge test to integration
